+++ app/src/main/AndroidManifest.xml
@@ -16,19 +16,19 @@
         android:theme="@style/Theme.PlaylistMaker2"
         tools:targetApi="31">
         <activity
-            android:name=".AudioplayerActivity"
+            android:name=".presentation.audioplayer.AudioplayerActivity"
             android:exported="false" />
         <activity
-            android:name=".SettingsActivity"
+            android:name=".presentation.settings.SettingsActivity"
             android:exported="false" />
         <activity
-            android:name=".SearchActivity"
+            android:name=".presentation.search.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".MedialibraryActivity"
+            android:name=".presentation.medialibrary.MedialibraryActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.PlaylistMaker2">
             <intent-filter>
