+++ app/src/main/java/com/marat/hvatit/playlistmaker2/Track.kt
@@ -1,0 +1,11 @@
+package com.marat.hvatit.playlistmaker2
+
+import android.os.Parcel
+import android.os.Parcelable
+
+data class Track(
+    val trackName: String,
+    val artistName: String,
+    val trackTime: String,
+    val artworkUrl100: String
+)
