+++ app/src/main/java/com/marat/hvatit/playlistmaker2/presentation/audioplayer/AudioplayerActivity.kt
@@ -1,4 +1,4 @@
-package com.marat.hvatit.playlistmaker2
+package com.marat.hvatit.playlistmaker2.presentation.audioplayer
 
 import android.content.Context
 import android.content.Intent
@@ -9,15 +9,14 @@ import android.os.Handler
 import android.os.Looper
 import android.util.Log
 import android.view.View
-import android.widget.Button
 import android.widget.ImageButton
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.core.view.isVisible
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.google.gson.Gson
-import com.marat.hvatit.playlistmaker2.service.AppleSong
+import com.marat.hvatit.playlistmaker2.data.network.AppleSong
+import com.marat.hvatit.playlistmaker2.R
 import java.text.SimpleDateFormat
 import java.util.Locale
