+++ app/src/main/java/com/marat/hvatit/playlistmaker2/presentation/medialibrary/MedialibraryActivity.kt
@@ -1,7 +1,8 @@
-package com.marat.hvatit.playlistmaker2
+package com.marat.hvatit.playlistmaker2.presentation.medialibrary
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.marat.hvatit.playlistmaker2.R
 
 class MedialibraryActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
