+++ app/src/main/res/layout/activity_audioplayer.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AudioplayerActivity">
+    tools:context=".presentation.audioplayer.AudioplayerActivity">
 
     <View
         android:id="@+id/back"
