+++ app/src/main/res/layout/activity_main.xml
@@ -2,7 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".MainActivity"
+    tools:context=".presentation.main.MainActivity"
     style="@style/activity_main_ll">
 
     <TextView
