+++ app/src/main/res/layout/activity_mainconstraint.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="#3772E7"
     android:padding="20dp"
-    tools:context=".MainActivity">
+    tools:context=".presentation.main.MainActivity">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guidelineone"
